Move rubric content to new component named RubricContent
