+++ apps/test/unit/templates/rubrics/RubricContentTest.jsx
@@ -3,9 +3,9 @@ import {expect} from '../../../util/reconfiguredChai';
 import {mount, shallow} from 'enzyme';
 import sinon from 'sinon';
 import HttpClient from '@cdo/apps/util/HttpClient';
-import RubricContainer from '@cdo/apps/templates/rubrics/RubricContainer';
+import RubricContent from '@cdo/apps/templates/rubrics/RubricContent';
 
-describe('RubricContainer', () => {
+describe('RubricContent', () => {
   const defaultRubric = {
     learningGoals: [
       {
@@ -42,7 +42,7 @@ describe('RubricContainer', () => {
 
   it('shows learning goals with correct props when viewing student work on assessment level', () => {
     const wrapper = shallow(
-      <RubricContainer
+      <RubricContent
         {...defaultProps}
         studentLevelInfo={{name: 'Grace Hopper', timeSpent: 706}}
       />
@@ -65,7 +65,7 @@ describe('RubricContainer', () => {
 
   it('shows learning goals with correct props when viewing student work on non assessment level', () => {
     const wrapper = shallow(
-      <RubricContainer
+      <RubricContent
         {...defaultProps}
         studentLevelInfo={{name: 'Grace Hopper', timeSpent: 706}}
         currentLevelName="non_assessment_level"
@@ -89,7 +89,7 @@ describe('RubricContainer', () => {
 
   it('shows learning goals with correct props when not viewing student work', () => {
     const wrapper = shallow(
-      <RubricContainer {...defaultProps} studentLevelInfo={null} />
+      <RubricContent {...defaultProps} studentLevelInfo={null} />
     );
     const renderedLearningGoals = wrapper.find('LearningGoal');
     expect(renderedLearningGoals).to.have.lengthOf(2);
@@ -110,7 +110,7 @@ describe('RubricContainer', () => {
   it('shows level title', () => {
     // mount is needed in order for text() to work
     const wrapper = mount(
-      <RubricContainer
+      <RubricContent
         {...defaultProps}
         studentLevelInfo={{
           name: 'Grace Hopper',
@@ -123,7 +123,7 @@ describe('RubricContainer', () => {
   it('shows student data if provided', () => {
     // mount is needed in order for text() to work
     const wrapper = mount(
-      <RubricContainer
+      <RubricContent
         {...defaultProps}
         studentLevelInfo={{
           name: 'Grace Hopper',
@@ -142,7 +142,7 @@ describe('RubricContainer', () => {
   it('handles missing student data', () => {
     // mount is needed in order for text() to work
     const wrapper = mount(
-      <RubricContainer
+      <RubricContent
         {...defaultProps}
         studentLevelInfo={{
           name: 'Grace Hopper',
@@ -158,7 +158,7 @@ describe('RubricContainer', () => {
   it('doesnt show student level data if not on level for evaluation', () => {
     // mount is needed in order for text() to work
     const wrapper = mount(
-      <RubricContainer
+      <RubricContent
         {...defaultProps}
         studentLevelInfo={{
           name: 'Grace Hopper',
@@ -174,7 +174,7 @@ describe('RubricContainer', () => {
 
   it('shows submit button if has student data and on level for evaluation', () => {
     const wrapper = shallow(
-      <RubricContainer
+      <RubricContent
         rubric={defaultRubric}
         currentLevelName="test_level"
         studentLevelInfo={{name: 'Grace Hopper'}}
@@ -186,7 +186,7 @@ describe('RubricContainer', () => {
 
   it('handles successful submit button click', async () => {
     const wrapper = shallow(
-      <RubricContainer
+      <RubricContent
         rubric={defaultRubric}
         teacherHasEnabledAi
         currentLevelName="test_level"
@@ -214,7 +214,7 @@ describe('RubricContainer', () => {
 
   it('handles error on submit button click', async () => {
     const wrapper = shallow(
-      <RubricContainer
+      <RubricContent
         rubric={defaultRubric}
         teacherHasEnabledAi
         studentLevelInfo={{name: 'Grace Hopper'}}
